% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-answer.R
\name{check_tenpuzzle}
\alias{check_tenpuzzle}
\alias{check_tenpuzzle_single}
\title{Check if an answer is a valid solution to a ten puzzle}
\usage{
check_tenpuzzle(expr, x, tgt = 10, useall = TRUE, intonly = FALSE,
  nonnegative = FALSE, nonzero = FALSE)

check_tenpuzzle_single(expr, x, tgt = 10, useall = TRUE, intonly = FALSE,
  nonnegative = FALSE, nonzero = FALSE)
}
\arguments{
\item{expr}{character of mathematical expression}

\item{x}{input integers}

\item{tgt}{target number}

\item{useall}{if \code{TRUE}, requires all numbers are used;
otherwise some number may remain unused}

\item{intonly}{if \code{TRUE}, requires all numbers are integers in
every step of calculation}

\item{nonnegative}{if \code{TRUE}, requires all numbers are nonnegative in
every step of calculation}

\item{nonzero}{if \code{TRUE}, requires all numbers are not zero in
every step of calculation}
}
\value{
A list of logical, with "reason" attribute for FALSE case
}
\description{
Check if an answer is a valid solution to a ten puzzle
}
\examples{
check_tenpuzzle("1+2+3+4", x=c(1,2,3,4))
check_tenpuzzle(c("2*4+3-1", "3*4+1-2"), x=c(1,2,3,4))
check_tenpuzzle(c("3*4-2", '(3-1)*2+4+2'), x=c(1,2,3,4))
check_tenpuzzle(c("4/2*3", '3/2*4'), x=c(2,3,4), tgt=6, intonly=TRUE)
}
