% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countdown.R
\name{countdown}
\alias{countdown}
\title{Solve Countdown's Number Quiz}
\usage{
countdown(x, tgt, findone = TRUE, useup = FALSE, intonly = TRUE,
  nonnegative = TRUE, nonzero = TRUE)
}
\arguments{
\item{x}{an integer vector}

\item{tgt}{an integer of target}

\item{findone}{if \code{TRUE}, search for an answer;
otherwise search exhaustively}

\item{useup}{if \code{TRUE}, requires all numbers are used;
otherwise some number may remain unused}

\item{intonly}{if \code{TRUE}, requires all numbers are integers in
every step of calculation; otherwise fractions may appear
in interim steps}

\item{nonnegative}{if \code{TRUE}, requires all numbers are nonnegative in
every step of calculation}

\item{nonzero}{if \code{TRUE}, requires all numbers are not zero in
every step of calculation}
}
\value{
a list of character vector of best possible answers and the values.
        if \code{findone} is selected and an exact match is found, then
        only one answer is returned.  If no exact match is found, then
        all closest possible solutions are returned.
}
\description{
Solve Countdown's Number Quiz
}
\examples{
countdown(c(75, 50, 2, 3, 8, 7), 812)
countdown(c(10, 5, 1, 7, 12, 62), 130)
countdown(c(8, 4, 3, 8, 87, 12), 724)
}
