% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eval_expr}
\alias{eval_expr}
\title{Convert mathematical expression into reverse polish notation}
\usage{
eval_expr(x)
}
\arguments{
\item{x}{character vector of mathematical expression}
}
\value{
a list
}
\description{
Convert mathematical expression into reverse polish notation
}
\examples{
eval_expr(c("1 + 2 + 3", "(4 - 1)*(-3 + 7) / 5", "(2*3 - 6)*(4 + 1/5)"))
}
