% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenpuzzle.R
\name{tenpuzzle}
\alias{tenpuzzle}
\title{Ten puzzle solver}
\usage{
tenpuzzle(x, tgt = 10, findone = TRUE, useall = TRUE, intonly = FALSE,
  nonnegative = FALSE, nonzero = FALSE)
}
\arguments{
\item{x}{an integer vector}

\item{tgt}{an integer of target}

\item{findone}{if \code{TRUE}, search for an answer;
otherwise search exhaustively}

\item{useall}{if \code{TRUE}, requires all numbers are used;
otherwise some number may remain unused}

\item{intonly}{if \code{TRUE}, requires all numbers are integers in
every step of calculation}

\item{nonnegative}{if \code{TRUE}, requires all numbers are nonnegative in
every step of calculation}

\item{nonzero}{if \code{TRUE}, requires all numbers are not zero in
every step of calculation}
}
\value{
a character vector of answers
}
\description{
Ten puzzle solver
}
\examples{
tenpuzzle(c(1, 1, 9, 9), 10)
tenpuzzle(c(1, 1, 9, 9), 10, findone=FALSE)
tenpuzzle(c(1, 1, 9, 9), 10, useall=FALSE)

# integer only
tenpuzzle(c(2, 6, 3), 9, findone=FALSE, intonly=FALSE)
tenpuzzle(c(2, 6, 3), 9, findone=FALSE, intonly=TRUE)
}
