% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{tenpuzzle_app}
\alias{tenpuzzle_app}
\alias{tenpuzzle_app_ui}
\alias{tenpuzzle_app_srv}
\title{Ten Puzzle Game on Shiny}
\usage{
tenpuzzle_app(...)

tenpuzzle_app_ui(id)

tenpuzzle_app_srv(input, output, session)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny]{runApp}}}
}
\description{
Ten Puzzle Game on Shiny
}
\examples{
tenpuzzle_app()
}
