% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensolver.R
\name{tenpuzzle_bf}
\alias{tenpuzzle_bf}
\title{Solve Ten Puzzle by Brute Force}
\usage{
tenpuzzle_bf(x, tgt = 10)
}
\arguments{
\item{x}{an integer vector}

\item{tgt}{an integer of target}
}
\value{
a character vector of all possible answers
}
\description{
Solve Ten Puzzle by Brute Force
}
\examples{
tenpuzzle_bf(c(1, 1, 9, 9), 10)
}
