% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhaustive.R
\name{tenpuzzle_exhaustive}
\alias{tenpuzzle_exhaustive}
\title{Solve Ten Puzzle by Brute Force}
\usage{
tenpuzzle_exhaustive(x, tgt = 10)
}
\arguments{
\item{x}{an integer vector}

\item{tgt}{an integer to target}
}
\value{
a character vector of all possible answers
}
\description{
Solve Ten Puzzle by Brute Force
}
\examples{
\dontrun{
tenpuzzle_exhaustive(c(1, 1, 9, 9), 10)
}
}
