% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{to_postfix}
\alias{to_postfix}
\title{Convert infix mathematical expression to postfix notation}
\usage{
to_postfix(x)
}
\arguments{
\item{x}{character vector of mathematical expression in the infix notation}
}
\value{
character vector of the postfix notation
}
\description{
Convert infix mathematical expression to postfix notation
}
\examples{
to_postfix(c("1 + 2 + 3", "(4 - 1)*(-3 + 7) / 5", "(2*3 - 6)*(4 + 1/5)"))
}
