% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-expr.R
\name{eval_expr}
\alias{eval_expr}
\title{Evaluate Mathematical Expressions}
\usage{
eval_expr(expr, valueonly = FALSE)
}
\arguments{
\item{expr}{character vector of mathematical expression}

\item{valueonly}{if \code{TRUE} only values are returned; otherwise
more information is returned}
}
\value{
if \code{valueonly} is \code{TRUE}, then a numeric vector of
calculation outcomes; otherwise, a list of four vectors of
the same size as the input:
\itemize{
\item{\code{value}: numeric vector of calculation outcomes}
\item{\code{intonly}: logical vector that
indicates that only integers appear in the intermediate steps}
\item{\code{nonzero}: logical vector that indicates that
zero does not appear in the intemediate steps}
\item{\code{nonnegative}: logical vector that indicates that
negative numbers do not appear in the intemediate steps}
}
}
\description{
Evaluate Mathematical Expressions
}
\examples{
eval_expr(c("1 + 2 + 3", "(4 - 1)*(-3 + 7) / 5", "(2*3 - 6)*(4 + 1/5)"))
}
