% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensolver.R
\name{tenpuzzle}
\alias{tenpuzzle}
\title{Ten puzzle solver}
\usage{
tenpuzzle(x, tgt = 10, findone = TRUE, useup = TRUE)
}
\arguments{
\item{x}{an integer vector}

\item{tgt}{an integer of target}

\item{findone}{if \code{TRUE}, search for an answer;
otherwise search exhaustively}

\item{useup}{if \code{TRUE}, requires all numbers are used;
otherwise some number may remain unused}
}
\value{
a character vector of answers
}
\description{
Ten puzzle solver
}
\examples{
tenpuzzle(c(1, 1, 9, 9), 10)
tenpuzzle(c(1, 1, 9, 9), 10, findone=FALSE)
tenpuzzle(c(1, 1, 9, 9), 10, useup=FALSE)

}
