% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{tenpuzzle_app}
\alias{tenpuzzle_app}
\title{Ten Puzzle Game on Shiny}
\usage{
tenpuzzle_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny]{runApp}}}
}
\description{
Ten Puzzle Game on Shiny
}
\examples{
\dontrun{
tenpuzzle_app()
}
}
